updated MSVC project files to use the new (generated) scripts/boost/include path for any boost includes.

+++ platform/win32/prep_boost_include_directories.sh
@@ -7,21 +7,22 @@ pushd "$( dirname "$0" )"
 
 cd ../../scripts
 
-# ditch the old collective: some files might have been removed or renamed (moved!) in the meantime!
+echo "Ditching the old collective: some files might have been removed or renamed (moved!) in the meantime!"
 
-rm -rf boost/include
+# rm -rf boost/include
 
-# no go and set it all up...
+echo "And now go and set it all up..."
 
 mkdir -p boost/include
 
 cd boost/include
 
-for f in $( find ../../../thirdparty/owemdjee/boost -type f \( -name '*.h' -o -name '*.hpp' \) | grep -e 'include\/' | sort ) ; do
+for f in $( find ../../../thirdparty/owemdjee/boost -type f \( -name '*.h' -o -name '*.hpp' \) | grep -e '\/include\/boost\/' | sort ) ; do
 	DSTFILE=$( echo $f | sed -E -e 's/^.*owemdjee//' -e 's/^.*include\///' )
-	echo "symlink: $DSTFILE --> $f"
-	mkdir -p $( dirname ./$DSTFILE )
-	cat > ./$DSTFILE << EOF
+	if ! test -f ./$DSTFILE ; then 
+		echo "symlink: $DSTFILE --> $f"
+		mkdir -p $( dirname ./$DSTFILE )
+		cat > ./$DSTFILE << EOF
 //
 // symlink-like file is generated using the platform/win32/prep_boost_include_directories.sh shell script.
 //
@@ -33,9 +34,10 @@ for f in $( find ../../../thirdparty/owemdjee/boost -type f \( -name '*.h' -o -n
 #include "$f"
 
 EOF
-	
+	fi
 done
 
+echo "Done!"
 
 popd
